Split state.js to eliminate circular dependencies in Extensions/common
state.js previously mixed pure config/predicate data (extConfig,
isMobile, isShorts, DEFAULT_SELECTORS, ...) with orchestration logic
that calls into buttons.js/bar.js/utils.js, while those files imported
the pure data back from state.js - a genuine four-file dependency
cycle that Rollup's stricter analysis flagged as warnings.

Extracts the pure pieces into two new leaf modules:
- dom.js: zero-dependency DOM/browser helpers (querySelector,
  getVideoId, getBrowser, localize, etc.)
- runtime-state.js: extConfig, storedData, and the page-detection
  predicates (isMobile, isShorts, isNewDesign, isRoundedDesign)

isLikesDisabled moves from state.js to buttons.js, since it inspects
button DOM state and was the other half of the bar.js <-> state.js
cycle (bar.js needed isLikesDisabled from state.js while state.js
needed createRateBar from bar.js).

state.js and utils.js re-export everything under their original names,
so no consumer outside Extensions/common needed to change - the
extension entry points, userscript entry point, and both userscript
override modules are untouched. The dependency graph is now a clean
DAG and Rollup no longer reports circular dependencies.
