+++ Extensions/common/dom.js
@@ -1,0 +1,152 @@
+function getBrowser() {
+  if (typeof chrome !== "undefined" && typeof chrome.runtime !== "undefined") {
+    return chrome;
+  } else if (typeof browser !== "undefined" && typeof browser.runtime !== "undefined") {
+    return browser;
+  } else {
+    console.log("browser is not supported");
+    return false;
+  }
+}
+
+function getVideoId(url) {
+  const urlObject = new URL(url);
+  const pathname = urlObject.pathname;
+  if (pathname.startsWith("/clip")) {
+    return (document.querySelector("meta[itemprop='videoId']") || document.querySelector("meta[itemprop='identifier']"))
+      .content;
+  } else {
+    if (pathname.startsWith("/shorts")) {
+      return pathname.slice(8);
+    }
+    return urlObject.searchParams.get("v");
+  }
+}
+
+function isInViewport(element) {
+  const rect = element.getBoundingClientRect();
+  const height = innerHeight || document.documentElement.clientHeight;
+  const width = innerWidth || document.documentElement.clientWidth;
+  return (
+    // When short (channel) is ignored, the element (like/dislike AND short itself) is
+    // hidden with a 0 DOMRect. In this case, consider it outside of Viewport
+    !(rect.top == 0 && rect.left == 0 && rect.bottom == 0 && rect.right == 0) &&
+    rect.top >= 0 &&
+    rect.left >= 0 &&
+    rect.bottom <= height &&
+    rect.right <= width
+  );
+}
+
+function getNumberFormatter(optionSelect) {
+  let userLocales;
+  if (document.documentElement.lang) {
+    userLocales = document.documentElement.lang;
+  } else if (navigator.language) {
+    userLocales = navigator.language;
+  } else {
+    try {
+      userLocales = new URL(
+        Array.from(document.querySelectorAll("head > link[rel='search']"))
+          ?.find((n) => n?.getAttribute("href")?.includes("?locale="))
+          ?.getAttribute("href"),
+      )?.searchParams?.get("locale");
+    } catch {
+      console.log("Cannot find browser locale. Use en as default for number formatting.");
+      userLocales = "en";
+    }
+  }
+
+  let formatterNotation;
+  let formatterCompactDisplay;
+  switch (optionSelect) {
+    case "compactLong":
+      formatterNotation = "compact";
+      formatterCompactDisplay = "long";
+      break;
+    case "standard":
+      formatterNotation = "standard";
+      formatterCompactDisplay = "short";
+      break;
+    case "compactShort":
+    default:
+      formatterNotation = "compact";
+      formatterCompactDisplay = "short";
+  }
+
+  return Intl.NumberFormat(userLocales, {
+    notation: formatterNotation,
+    compactDisplay: formatterCompactDisplay,
+  });
+}
+
+function localize(localeString, substitutions) {
+  try {
+    if (typeof chrome !== "undefined" && chrome?.i18n?.getMessage) {
+      const args = substitutions === undefined ? [localeString] : [localeString, substitutions];
+      const message = chrome.i18n.getMessage(...args);
+      if (message) {
+        return message;
+      }
+    }
+  } catch (error) {
+    console.warn("Localization lookup failed for", localeString, error);
+  }
+
+  if (Array.isArray(substitutions)) {
+    return substitutions.join(" ");
+  }
+
+  if (substitutions != null) {
+    return `${substitutions}`;
+  }
+
+  return localeString;
+}
+
+function querySelector(selectors, element) {
+  let result;
+  for (const selector of Array.isArray(selectors) ? selectors : [selectors]) {
+    if (!selector) continue;
+    result = (element ?? document).querySelector(selector);
+    if (result !== null) {
+      return result;
+    }
+  }
+}
+
+function querySelectorAll(selectors) {
+  let result;
+  for (const selector of Array.isArray(selectors) ? selectors : [selectors]) {
+    if (!selector) continue;
+    result = document.querySelectorAll(selector);
+    if (result.length !== 0) {
+      return result;
+    }
+  }
+  return result ?? document.querySelectorAll("__ryd-missing-selector__");
+}
+
+function createObserver(options, callback) {
+  const observerWrapper = new Object();
+  observerWrapper.options = options;
+  observerWrapper.observer = new MutationObserver(callback);
+  observerWrapper.observe = function (element) {
+    this.observer.observe(element, this.options);
+  };
+  observerWrapper.disconnect = function () {
+    this.observer.disconnect();
+  };
+  return observerWrapper;
+}
+
+export {
+  getBrowser,
+  getVideoId,
+  isInViewport,
+  getNumberFormatter,
+  localize,
+  querySelector,
+  querySelectorAll,
+  createObserver,
+};
